Update Dockerfile - convert ggml to gguf and do things to support that
The ggml file needs to be converted to gguf. To support this, python, pip, and numpy need to be installed. Install them, run conversion, then uninstall them & remove ggml file in an attempt to keep file size down. Also, this stuff is a moving target right now so pin llamma cpp to specific commit.